services/visionsvc/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
transformers
torch --index-url https://download.pytorch.org/whl/cpu